requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
pydantic==2.12.3
newspaper3k==0.2.8
transformers==4.57.1
torch==2.8.0
python-multipart==0.0.20   # For form handling, optional
requests==2.32.0  # Newspaper3k dependency
nltk==3.9.2
spacy==3.8.7
tqdm==4.67.1
lxml[html_clean]